requirements.txt
antlr4-python3-runtime==4.13.0
z3-solver==4.12.2.0
typing-extensions==4.7.1